add entities test case
